requirements.txt
absl-py==0.9.0
alembic==0.9.9
aniso8601==8.0.0
astunparse==1.6.3
bcrypt==3.1.7
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
flask-marshmallow==0.13.0
Flask-Migrate==2.1.1
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.11.3
idna==2.6
importlib-metadata==1.7.0
itsdangerous==0.24
Jinja2==2.10
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy==1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.7.0
matplotlib==3.3.0
mysqlclient==2.0.1
numpy==1.18.5
oauth2==1.9.0.post1
oauthlib==2.0.7
opt-einsum==3.3.0
Pillow==7.2.0
plotly==4.9.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.4.2
pyparsing==2.4.7
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==0.8.0
retrying==1.3.3
rsa==4.6
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.22
URLObject==2.4.3
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.1
zipp==3.1.0
